Try to fix: polars.exceptions.SchemaError: type Int64 is incompatible with expected type Int32

+++ src/process_data.py
@@ -40,7 +40,7 @@ def main(experiment_path: Path):
                 pl.lit(env).alias("env"),
                 pl.lit(group).alias("group"),
                 pl.lit(alg).alias("alg"),
-                pl.lit(aperture).cast(pl.Int32).alias("aperture"),
+                pl.lit(aperture).cast(pl.Int64).alias("aperture"),
             )
             print(df)
             dfs.append(df)
